+++ src/create/svelte/sveltekit.js
@@ -1,21 +1,25 @@
-import * as svelteKit from "./sveltekit/general.js"
+import * as general from "./sveltekit/general.js"
+import * as extra from "./sveltekit/extra.js"
 import * as templates from "./sveltekit/templates.js"
+import * as typescript from "./sveltekit/typescript.js"
 const withDep = [ "database" ]
 
 export default function sveltekit ( files, options ) {
-	let pkgIndex = files.findIndex(( file ) => file.name == "package.json")
-	let pkg = files[pkgIndex].content
+	const pkgIndex = files.findIndex(({ name }) => name == "package.json")
 
-	pkg.scripts = {
-		dev: "svelte-kit dev",
-		build: "svelte-kit build --verbose",
-		preview: "svelte-kit preview"
-	}
-
-	pkg.devDependencies = {
-		[`@sveltejs/adapter-${options.adapter}`]: "next",
-		"@sveltejs/kit": "next",
-		"svelte": "^3.38.2"
+	let pkg = {
+		...files[pkgIndex].content,
+		scripts: {
+			dev: "svelte-kit dev",
+			build: "svelte-kit build --verbose",
+			preview: "svelte-kit preview"
+		},
+		devDependencies: {
+			[`@sveltejs/adapter-${options.adapter}`]: "next",
+			"@sveltejs/kit": "next",
+			"svelte": "^3.38.2"
+		},
+		type: "module"
 	}
 
 	if (options.typescript) {
@@ -26,51 +30,43 @@ export default function sveltekit ( files, options ) {
 		}
 	}
 
-	delete pkg.main
-
 	if (options.extra.some(( extra ) => withDep.includes(extra)) || options.fonts.length) pkg.dependencies = {}
 
 	options.fonts.forEach(( font ) => {
 		pkg.dependencies[`@fontsource/${font}`] = "^4.4.5"
 	})
 
-	pkg.type = "module"
-
 	files[pkgIndex].content = pkg
 
-	let gitignoreIndex = files.findIndex(( file ) => file.name == ".gitignore")
-	files[gitignoreIndex].content += svelteKit.gitignore
+	const gitignoreIndex = files.findIndex(({ name }) => name == ".gitignore")
+	files[gitignoreIndex].content += general.gitignore
 
 	files = files.concat([
 		{
 			name: "svelte.config.js",
-			content: svelteKit.config(options.typescript, options.adapter, options.extra)
-		},
-		{
-			name: "todo",
-			content: svelteKit.todo
+			content: general.config(options.typescript, options.adapter, options.extra)
 		},
 		{
 			name: "src",
 			files: [
 				{
 					name: "app.css",
-					content: svelteKit.css(options.fonts)
+					content: general.css(options.fonts)
 				},
 				{
 					name: "app.html",
-					content: svelteKit.app
+					content: general.app
 				},
 				{
 					name: "routes",
 					files: [
 						{
 							name: "__layout.svelte",
-							content: svelteKit.__layout(options.typescript)
+							content: general.__layout(options.typescript)
 						},
 						{
 							name: "index.svelte",
-							content: svelteKit.index(options.typescript)
+							content: general.index(options.typescript)
 						}
 					]
 				},
@@ -85,48 +81,48 @@ export default function sveltekit ( files, options ) {
 			files: [
 				{
 					name: "robots.txt",
-					content: svelteKit.robots
+					content: general.robots
 				}
 			]
 		}
 	])
 
 	if (options.typescript) {
-		let srcIndex = files.findIndex(( file ) => file.name == "src")
+		const srcIndex = files.findIndex(({ name }) => name == "src")
 
 		files[srcIndex].files.push({
 			name: "global.d.ts",
-			content: svelteKit.ts.global
+			content: typescript.global
 		})
 
 		files.push({
 			name: "tsconfig.json",
-			content: svelteKit.ts.tsconfig
+			content: typescript.tsconfig
 		})
 	}
 
 	if (options.templates.length) {
-		let srcIndex = files.findIndex(( file ) => file.name == "src")
-		let srcLibIndex = files[srcIndex].files.findIndex(( file ) => file.name == "lib")
-		let srcRoutesIndex = files[srcIndex].files.findIndex(( file ) => file.name == "routes")
+		const srcIndex = files.findIndex(({ name }) => name == "src")
+		const srcLibIndex = files[srcIndex].files.findIndex(({ name }) => name == "lib")
+		const srcRoutesIndex = files[srcIndex].files.findIndex(({ name }) => name == "routes")
 
 		options.templates.forEach(( template ) => {
 			switch (template) {
 				case "__error": {
 					files[srcIndex].files[srcRoutesIndex].files.push({
 						name: "__error.svelte",
-						content: templates.__error
+						content: templates.__error(options.typescript)
 					})
 					break
 				}
 				case "message": {
 					files[srcIndex].files[srcLibIndex].files.push({
 						name: "message.svelte",
-						content: templates.message.svelte
+						content: templates.message.svelte(options.typescript)
 					},
 					{
-						name: "message.js",
-						content: templates.message.js
+						name: !options.typescript ? "message.js" : "message.ts",
+						content: !options.typescript ? templates.message.js : templates.message.ts
 					})
 				}
 			}
@@ -134,42 +130,44 @@ export default function sveltekit ( files, options ) {
 	}
 
 	if (options.extra.length) {
-		let srcIndex = files.findIndex(( file ) => file.name == "src")
-		let srcLibIndex = files[srcIndex].files.findIndex(( file ) => file.name == "lib")
-		let pkg = files.findIndex(( file ) => file.name == "package.json")
+		const srcIndex = files.findIndex(({ name }) => name == "src")
+		const srcLibIndex = files[srcIndex].files.findIndex(({ name }) => name == "lib")
+		const pkg = files.findIndex(({ name }) => name == "package.json")
 
-		options.extra.forEach(( extra ) => {
-			switch (extra) {
+		options.extra.forEach(( ex ) => {
+			switch (ex) {
 				case "database": {
 					switch (options.database) {
 						case "mongodb": {
-							files[pkg].content.dependencies["mongodb"] = "^3.6.9"
+							files[pkg].content.dependencies.mongodb = "^3.6.9"
 
 							files[srcIndex].files[srcLibIndex].files.push({
-								name: "mongodb.js",
-								content: svelteKit.extra.mongodb
+								name: !options.typescript ? "mongodb.js" : "mongodb.ts",
+								content: !options.typescript ? extra.mongodb.js : extra.mongodb.ts
 							})
 
+							if (options.typescript) files[pkg].content.devDependencies["@types/mongodb"] = "^3.6.18"
 							break
 						}
 						case "mysql": {
-							files[pkg].content.dependencies["mysql"] = "^2.18.1"
+							files[pkg].content.dependencies.mysql = "^2.18.1"
 
 							files[srcIndex].files[srcLibIndex].files.push({
-								name: "mysql.js",
-								content: svelteKit.extra.mysql
+								name: !options.typescript ? "mysql.js" : "mysql.ts",
+								content: !options.typescript ? extra.mysql.js : extra.mysql.ts
 							})
 
+							if (options.typescript) files[pkg].content.devDependencies["@types/mysql"] = "^2.15.18"
 							break
 						}
 					}
-
 					break
 				}
 				case "preprocess": {
 					if (!options.typescript) {
 						files[pkg].content.devDependencies["svelte-preprocess"] = "^4.7.3"
 					}
+					break
 				}
 			}
 		})