+++ src/create/svelte/sveltekit/general.js
@@ -44,66 +44,16 @@ export const css = ( fonts ) => [
 
 export const config = ( typescript, adapter, extra ) => [
 	`import ${adapter == "static" ? "adapter" : adapter} from \"@sveltejs/adapter-${adapter}\"`,
-	typescript || extra.includes("preoprocess") ? "import sveltePreprocess from \"svelte-preprocess\"" : null,
+	typescript || extra.includes("preprocess") ? "import sveltePreprocess from \"svelte-preprocess\"" : null,
 	"",
 	"export default {",
-	typescript || extra.includes("preoprocess") ? "\tpreprocess: sveltePreprocess(),\n" : null,
+	typescript || extra.includes("preprocess") ? "\tpreprocess: sveltePreprocess()," : null,
 	"\tkit: {",
 	`\t\tadapter: ${adapter == "static" ? "adapter" : adapter}()`,
 	"\t}",
 	"}",
 	""
 ].filter(( line ) => line != null).join("\n")
-
-export const extra = {
-	mongodb: [
-		"import mongodb from \"mongodb\"",
-		"const { MongoClient, ObjectID } = mongodb",
-		"",
-		"let client = null",
-		"let db = null",
-		"",
-		"export async function init () {",
-		"\tif (!client) {",
-		"\t\tclient = await MongoClient.connect(\"mongodb://localhost:27017/\", { useNewUrlParser: true, useUnifiedTopology: true })",
-		"\t\tdb = client.db(\"name\")",
-		"\t}",
-		"",
-		"\treturn { db, client }",
-		"}",
-		"",
-		"export { ObjectID }",
-		""
-	].join("\n"),
-	mysql: [
-		"import { createConnection } from \"mysql\"",
-		"",
-		"const options = {",
-		"\thost: \"host\",",
-		"\tuser: \"user\",",
-		"\tpassword: \"password\",",
-		"\tdatabase: \"database\"",
-		"}",
-		"",
-		"let db = null",
-		"",
-		"export async function init () {",
-		"\tif (!db) {",
-		"\t\tdb = createConnection(options)",
-		"\t\tawait new Promise(( resolve, reject ) => db.connect((( err ) => err ? reject(err) : resolve())))",
-		"\t}",
-		"",
-		"\treturn { db }",
-		"}",
-		"",
-		"export function query ( db, query ) {",
-		"\treturn new Promise(( resolve ) => {",
-		"\t\tdb.query(query, ( error, data ) => error ? resolve(null) : resolve(data))",
-		"\t})",
-		"}",
-		""
-	].join("\n")
-}
 
 export const gitignore = [
 	"",
@@ -114,7 +64,7 @@ export const gitignore = [
 ].join("\n")
 
 export const index = ( typescript ) => [
-	`<script context=\"module\"${typescript ? " lang=\"ts\"" : ""}>`,
+	`<script${typescript ? " lang=\"ts\"" : ""} context=\"module\">`,
 	"\texport const prerender = true",
 	"</script>",
 	"",
@@ -129,68 +79,3 @@ export const robots = [
 	"Disallow:",
 	""
 ].join("\n")
-
-export const todo = [
-	"[ content ]",
-	"[ ui ]",
-	"[ bugs ]",
-	"[ misc ]",
-	""
-].join("\n")
-
-export const ts = {
-	tsconfig: [
-		"{",
-		"\t\"compilerOptions\": {",
-		"\t\t/* - project options - */",
-		"\t\t\"allowJs\": true,",
-		"\t\t\"checkJs\": true,",
-		"\t\t\"isolatedModules\": true,",
-		"\t\t\"lib\": [\"es2020\"],",
-		"\t\t\"module\": \"es2020\",",
-		"\t\t\"removeComments\": true,",
-		"\t\t\"sourceMap\": true,",
-		"\t\t\"target\": \"es2019\",",
-		"",
-		"\t\t/* - strict checks - */",
-		"\t\t\"alwaysStrict\": true,",
-		"\t\t\"noImplicitAny\": true,",
-		"\t\t\"noImplicitThis\": true,",
-		"\t\t\"strict\": true,",
-		"\t\t\"strictBindCallApply\": true,",
-		"\t\t\"strictFunctionTypes\": true,",
-		"\t\t\"strictNullChecks\": true,",
-		"\t\t\"strictPropertyInitialization\": true,",
-		"",
-		"\t\t/* - linter checks - */",
-		"\t\t\"noImplicitReturns\": true,",
-		"\t\t\"noFallthroughCasesInSwitch\": true,",
-		"\t\t\"noUncheckedIndexedAccess\": true,",
-		"\t\t\"noUnusedLocals\": true,",
-		"\t\t\"noUnusedParameters\": true,",
-		"",
-		"\t\t/* - module resolution - */",
-		"\t\t\"baseUrl\": \".\",",
-		"\t\t\"esModuleInterop\": true,",
-		"\t\t\"moduleResolution\": \"node\",",
-		"\t\t\"paths\": {",
-		"\t\t\t\"lib/*\": [\"src/lib/*\"]",
-		"\t\t},",
-		"",
-		"\t\t/* - advanced - */",
-		"\t\t\"forceConsistentCasingInFileNames\": true,",
-		"\t\t\"importsNotUsedAsValues\": \"error\",",
-		"\t\t\"resolveJsonModule\": true,",
-		"\t\t\"skipLibCheck\": true,",
-		"\t},",
-		"\t\"include\": [\"src/**/*.d.ts\", \"src/**/*.js\", \"src/**/*.ts\", \"src/**/*.svelte\"]",
-		"}",
-		""
-	].join("\n"),
-	global: [
-		"/// <reference types=\"@sveltejs/kit\" />",
-		"/// <reference types=\"svelte\" />",
-		"/// <reference types=\"vite/client\" />",
-		""
-	].join("\n")
-}